% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_bool}
\alias{replace_bool}
\title{Replace \code{true} and \code{false} values with \code{1} and \code{0}, respectively}
\usage{
replace_bool(suso_expr)
}
\arguments{
\item{suso_expr}{Character vector. SuSo enablement or validation condition.}
}
\value{
Character vector. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace \code{true} and \code{false} values with \code{1} and \code{0}, respectively
}
