% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contains.R
\name{replace_Contains}
\alias{replace_Contains}
\title{Replace all of SuSo's Contains calls with Stata equivalent}
\usage{
replace_Contains(suso_expr)
}
\arguments{
\item{suso_expr}{Character vector. SuSo enablement or validation condition.}
}
\value{
Character vector. Expression where SuSo content has been replaced
with Stata content.
}
\description{
Replace all of SuSo's Contains calls with Stata equivalent
}
