% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contains.R
\name{replace_Contains_yn}
\alias{replace_Contains_yn}
\title{Replace all SuSo's Yes.Contains and No.Contains calls with Stata equivalent}
\usage{
replace_Contains_yn(suso_expr)
}
\arguments{
\item{suso_expr}{Character. SuSo enablement or validation condition.}
}
\value{
Character. Expression where SuSo content has been replaced with
Stata content.
}
\description{
Replace all SuSo's Yes.Contains and No.Contains calls with Stata equivalent
}
