% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contains.R
\name{replace_ContainsAll}
\alias{replace_ContainsAll}
\title{Replace all of SuSo's ContainsAll call with Stata equivalent}
\usage{
replace_ContainsAll(suso_expr)
}
\arguments{
\item{suso_expr}{Character vector. SuSo enablement or validation condition.}
}
\value{
Character vector. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace all of SuSo's ContainsAll call with Stata equivalent
}
