% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contains.R
\name{replace_ContainsAny}
\alias{replace_ContainsAny}
\title{Replace all of SuSo's ContainsAny calls with the Stata equivalent(s)}
\usage{
replace_ContainsAny(suso_expr)
}
\arguments{
\item{suso_expr}{Character. SuSo enablement or validation condition.}
}
\value{
Character. Expression where SuSo content has been replaced with
Stata content.
}
\description{
Replace all of SuSo's ContainsAny calls with the Stata equivalent(s)
}
