% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InList_InRange.R
\name{replace_InList}
\alias{replace_InList}
\title{Replace all of SuSo's InList with equivalent Stata inlist}
\usage{
replace_InList(suso_expr)
}
\arguments{
\item{suso_expr}{Character. SuSo enablement or validation condition.}
}
\value{
Character. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace all of SuSo's InList with equivalent Stata inlist
}
