% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InList_InRange.R
\name{replace_InRange}
\alias{replace_InRange}
\title{Replace all of SuSo's InRange with equivalent Stata inrange}
\usage{
replace_InRange(suso_expr)
}
\arguments{
\item{suso_expr}{Character vector. SuSo enablement or validation condition.}
}
\value{
Character vector. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace all of SuSo's InRange with equivalent Stata inrange
}
