% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contains.R
\name{replace_one_Contains}
\alias{replace_one_Contains}
\title{Replace SuSo's Contains with Stata equivalent}
\usage{
replace_one_Contains(suso_expr)
}
\arguments{
\item{suso_expr}{Character. SuSo enablement or validation condition.}
}
\value{
Character. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace SuSo's Contains with Stata equivalent
}
