% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contains.R
\name{replace_one_Contains_yn}
\alias{replace_one_Contains_yn}
\title{Replace SuSo's Yes.Contains and No.Contains}
\usage{
replace_one_Contains_yn(suso_expr)
}
\arguments{
\item{suso_expr}{Character. SuSo enablement or validation condition.}
}
\value{
Character. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace SuSo's Yes.Contains and No.Contains
}
