% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InList_InRange.R
\name{replace_one_InRange}
\alias{replace_one_InRange}
\title{Replace SuSo's InRange with equivalent Stata inrange}
\usage{
replace_one_InRange(suso_expr)
}
\arguments{
\item{suso_expr}{Character. SuSo enablement or validation condition.}
}
\value{
Character. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace SuSo's InRange with equivalent Stata inrange
}
