% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_rowcode_with}
\alias{replace_rowcode_with}
\title{Replace rowcode or roster variable references with row ID variable}
\usage{
replace_rowcode_with(suso_expr, rowcode = "@rowcode", replacement)
}
\arguments{
\item{suso_expr}{Character vector. SuSo enablement or validation condition.}

\item{rowcode}{Character. \code{rowcode} mention like \verb{@rowcode} or '
\verb{@roster_var_name}}

\item{replacement}{Character. Row ID variable replacement
(e.g., \code{members__id}).}
}
\value{
Character vector. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace rowcode or roster variable references with row ID variable
}
