% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_and}
\alias{replace_and}
\title{Replace SuSo \code{&&} with Stata \code{&}}
\usage{
replace_and(suso_expr)
}
\arguments{
\item{suso_expr}{Character vector. SuSo enablement or validation condition.}
}
\value{
Character vector. If pattern found, modified expression.
Otherwise, unmodified expression.
}
\description{
Replace SuSo \code{&&} with Stata \code{&}
}
