% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_IsAnswered}
\alias{replace_IsAnswered}
\title{Replace SuSo IsAnswered with Stata mi}
\usage{
replace_IsAnswered(suso_expr)
}
\arguments{
\item{suso_expr}{Character. SuSo enablement or validation condition.}
}
\description{
Replace SuSo IsAnswered with Stata mi
}
\details{
First, replace all negated IsAnswered calls
(i.e., \code{!IsAnswered()}). Then, replace all non-negated calls
(i.e., \code{IsAnswered()}). In that way, negated calls will be handled properly.
}
